Refactor chatbox page and chat history components
